requirements.txt
azure-functions
azure-identity
httpx
